--- src/components/SearchBar/index.ts
@@ -1,1 +1,0 @@
-export { SearchBar } from './SearchBar.tsx';
